Added GOAP base
feat: GOAP WorldState added.
A WorldState describes the state of the "world". These states are used by the planner to find which actions can be executed on the current world state.

feat: GOAP Planner added.
The planner uses A* to find the best course of actions.

feat: GOAP Action added,
An action represents a task the A.I. can execute. The planner chains these actions together to formulate a plan

feat: GOAP Node added.
A node is part of the "graph" of actions the planner will traverse.