Improved old KillEnemy actions.
impr: KillEnemyPistol and KillEnemyShotgun will now always turn the agent along the smallest angle to face the zombie. Previously, whenever the desired orientation was at -175 degrees and the agent's current orienatation was at 170 degrees, the agent would rotate all the way to the right because de difference between desired and current was negative.